requirements.txt
beautifulsoup4
requests
mechanize
html5lib
argparse
tqdm
tinydb
google-api-python-client
google-auth-httplib2
google-auth-oauthlib